fix: update offset type in BlockMeta and SstableWriter and use file.stream_position

+++ engine/src/sstable/index.rs
@@ -3,12 +3,12 @@ use bitcode::Encode;
 #[derive(Debug, Encode)]
 pub struct BlockMeta {
     len: usize,
-    offset: usize,
+    offset: u64,
     last_key: Vec<u8>,
 }
 
 impl BlockMeta {
-    pub fn new(len: usize, offset: usize, last_key: Vec<u8>) -> Self {
+    pub fn new(len: usize, offset: u64, last_key: Vec<u8>) -> Self {
         Self {
             len,
             offset,
